requirements.txt
altgraph==0.17.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
cryptography==41.0.4
Markdown==3.4.4
passlib==1.7.4
pefile==2023.2.7
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQt5-stubs==5.15.6.0
pywin32-ctypes==0.2.2
